pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "khabaroff-public-feedback-bot"
version = "0.1.0"
description = "Telegram bot for collecting and generating public feedback"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "aiogram>=3.4.1,<4.0.0",
  "aiohttp>=3.9.0,<4.0.0",
  "PyYAML>=6.0.1,<7.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0,<9.0.0",
  "pytest-asyncio>=0.23.0,<1.0.0",
]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.setuptools]
packages = ["bot"]
